This PR adds the ability for QNodes and quantum tapes to cache previous evaluations:
```python
from pennylane.beta.tapes import qnode
from pennylane.beta.queuing import expval
import time

dev = qml.device("default.qubit", wires=2)

@qnode(dev, caching=10)  # cache up to 10 evaluations
def qfunc(x):
    qml.RX(x, wires=0)
    qml.RX(0.3, wires=1)
    qml.CNOT(wires=[0, 1])
    return expval(qml.PauliZ(1))

for i in range(100):
    qfunc(0.1)
# Even with 100 calls, the internal execute_device() method
# of the quantum tape has only been called once
```

This is a replacement of https://github.com/PennyLaneAI/pennylane/pull/776 to work with the new QTape core.

### Things to note

- Caching works with both the underlying tape and the QNode.
- Caching is off by default, so it should not break anything
- QNode caching _implicitly_ assumes that the QNode is immutable. This is communicated to the user via the docstrings and also a warning when caching is turned on. To make this non-implicit, we could either (a) support immutability explicitly, and only allow for caching in that case, (b) associate different caches to different circuits by using the hash of the circuit. Option (b) sounds best but we do not currently support circuit hashing in the new core. This could be added, but is perhaps better be done more carefully using openqasm serialization.
- Caching still takes place if the user is returning samples or if operating in non-analytic mode. It isn't hard to prevent caching in these cases (and is done in https://github.com/PennyLaneAI/pennylane/pull/776), but I ended up being on the fence about whether to. It seems too implicit to turn off caching - indeed if a user turns on caching then they should get what they are expecting (cached results). Is there a case people can think of as a counterargument?

## Benchmarking

Let's take a look to see if caching is working. We'll consider a 6-qubit circuit composed of `StronglyEntanglingLayers` of depth 6 (similarly to https://github.com/PennyLaneAI/pennylane/pull/776). We will also look at benchmarking both the quantum tape and the QNode in two settings: evaluation (with 500 repetitions) and gradient evaluation (with 20 repetitions), where each repetition is with the same input parameters.

### Tape

Without caching:
Evaluate - 5.36 seconds
Gradient - 24.67 seconds

With caching:
Evaluate - 0.18 seconds
Gradient - 1.43 seconds

### QNode

Without caching:
Evaluate - 9.79 seconds
Gradient - 69.20 seconds

With caching:
Evaluate - 3.92 seconds
Gradient - 3.98 seconds

### Summary

Caching is helping us! Also, we can see that using the QNode has an additional overhead, e.g. from methods such as `construct()`.