@trbromley 2020-09-21
Closing to remake as a PR to `master`: https://github.com/PennyLaneAI/pennylane/pull/817